feat: 나의알림 탭 삭제하기 버튼 & 함께 기다리는 사람 추가
