+++ src/components/product/delete-product/index.css.ts
@@ -1,0 +1,16 @@
+import { style } from '@vanilla-extract/css';
+
+export const deleteBtn = style({
+  all: 'unset',
+  width: '7.7rem',
+  height: '3.3rem',
+  fontSize: '1.4rem',
+  fontWeight: '600',
+  textAlign: 'center',
+  marginTop: '0.6rem',
+  color: '#101010',
+  backgroundColor: '#FAFAFA',
+  border: '0.1rem solid #EEEEEE',
+  borderRadius: '0.4rem',
+  cursor: 'pointer',
+});
