+++ src/components/product/my-basket-card/index.tsx
@@ -1,11 +1,10 @@
 /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 import { useRouter } from 'next/navigation';
-import { useDeleteBasket } from '@/hooks/mutations/use-delete-basket';
-import { useModalStore } from '@/store/use-modal-store';
 import { type MyBasket } from '@/types/basket';
 
 import { BasketBadge } from '../basket-badge';
+import DeleteProductBtn from '../delete-product';
 import ProductThumbnailImage from '../product-thumbnail';
 import WaitingCounts from '../waiting-counts';
 import * as styles from './index.css';
@@ -27,20 +26,6 @@ const MyBasketCard = ({
 }: MyBasketProps) => {
   const description = `${brand}${firstOption ? ` / ${firstOption}` : ''}${secondOption ? ` / ${secondOption}` : ''}${thirdOption ? ` / ${thirdOption}` : ''}`;
 
-  const { mutate: deleteBasket } = useDeleteBasket();
-
-  const onOpen = useModalStore(state => state.onOpen);
-
-  const handleDeleteBtn = () => {
-    onOpen({
-      modalType: 'confirm',
-      title: '상품을 삭제하시겠습니까?',
-      description: name,
-      onClick: () => deleteBasket(id),
-      mainBtnText: '삭제',
-    });
-  };
-
   const router = useRouter();
 
   const openProductDetail = () => {
@@ -57,9 +42,7 @@ const MyBasketCard = ({
         </p>
         <p className={styles.options}>{description}</p>
         <WaitingCounts counts={waitingCount} />
-        <button className={styles.deleteBtn} onClick={handleDeleteBtn}>
-          삭제하기
-        </button>
+        <DeleteProductBtn id={id} name={name} />
       </div>
     </div>
   );
